스플래시 스크린 분리
